+++ kaithem/src/chandler/resource_type.py
@@ -21,7 +21,7 @@ def set_save_cb(c: WebChandlerConsole.WebConsole, module: str, resource: str):
 
 
 class ConfigType(modules_state.ResourceType):
-    def blurb(self, module, resource, object):
+    def blurb(self, module, resource, data):
         return f"""
         <div class="tool-bar">
             <a href="/chandler/editor/{module}:{resource}">
@@ -38,61 +38,61 @@ class ConfigType(modules_state.ResourceType):
         </div>
         """
 
-    def onload(self, module, resourcename, value):
-        value = copy.deepcopy(value)
-        x = entries.pop((module, resourcename), None)
-        entries[module, resourcename] = WebChandlerConsole.WebConsole(f"{module}:{resourcename}")
-        set_save_cb(entries[module, resourcename], module, resourcename)
-        core.boards[f"{module}:{resourcename}"] = entries[module, resourcename]
+    def on_load(self, module, resource, data):
+        data = copy.deepcopy(data)
+        x = entries.pop((module, resource), None)
+        entries[module, resource] = WebChandlerConsole.WebConsole(f"{module}:{resource}")
+        set_save_cb(entries[module, resource], module, resource)
+        core.boards[f"{module}:{resource}"] = entries[module, resource]
 
         if x:
             x.close()
 
-        core.boards[f"{module}:{resourcename}"].setup(value.get("project", {}))
+        core.boards[f"{module}:{resource}"].setup(data.get("project", {}))
 
-    def onmove(self, module, resource, toModule, toResource, resourceobj):
+    def on_move(self, module, resource, to_module, to_resource, data):
         x = entries.pop((module, resource), None)
         if x:
-            entries[toModule, toResource] = x
+            entries[to_module, to_resource] = x
 
         b = core.boards.pop(f"{module}:{resource}", None)
 
         if b:
-            b = core.boards[f"{toModule}:{toResource}"] = b
+            b = core.boards[f"{to_module}:{to_resource}"] = b
 
-        set_save_cb(entries[toModule, toResource], toModule, toResource)
+        set_save_cb(entries[to_module, to_resource], to_module, to_resource)
 
-    def onupdate(self, module, resource, obj):
-        self.onload(module, resource, obj)
+    def on_update(self, module, resource, data):
+        self.on_load(module, resource, data)
 
-    def ondelete(self, module, name, value):
+    def on_delete(self, module, name, data):
         entries[module, name].close()
         core.boards.pop(f"{module}:{name}", None)
 
         del entries[module, name]
 
-    def oncreaterequest(self, module, name, kwargs):
+    def on_create_request(self, module, resource, kwargs):
         d = {"resource_type": self.type}
         return d
 
-    def onupdaterequest(self, module, resource, resourceobj, kwargs):
-        d = resourceobj
+    def on_update_request(self, module, resource, data, kwargs):
+        d = data
         kwargs.pop("name", None)
         kwargs.pop("Save", None)
         return d
 
-    def createpage(self, module, path):
+    def create_page(self, module, path):
         d = dialogs.SimpleDialog("New Chandler Board")
         d.text_input("name", title="Resource Name")
 
         d.submit_button("Save")
         return d.render(self.get_create_target(module, path))
 
-    def editpage(self, module, name, value):
+    def edit_page(self, module, resource, data):
         d = dialogs.SimpleDialog("Editing Config")
         d.text("Edit the board in the chandler UI")
 
-        return d.render(self.get_update_target(module, name))
+        return d.render(self.get_update_target(module, resource))
 
     def flush_unsaved(self, module, resource):
         entries[module, resource].check_autosave()