+++ kaithem/src/plugins/CorePluginUserPageResources/fileserver.py
@@ -40,29 +40,29 @@ class FileServerType(modules_state.ResourceType):
     def blurb(self, m, r, value):
         return f'<a href="/pages/{url(m)}/{quote(r)}">Browse</a>'
 
-    def onload(self, module: str, resourcename: str, value: modules_state.ResourceDictType):
+    def on_load(self, module: str, resourcename: str, value: modules_state.ResourceDictType):
         by_module_resource[module, resourcename] = ServerObj(module, resourcename, value)
         if lookup:
             lookup.invalidate_cache()
 
-    def onmove(self, module, resource, toModule, toResource, resourceobj):
+    def on_move(self, module, resource, to_module, to_resource, resourceobj):
         x = by_module_resource.pop((module, resource), None)
         if x:
-            by_module_resource[toModule, toResource] = x
+            by_module_resource[to_module, to_resource] = x
 
         if lookup:
             lookup.invalidate_cache()
 
-    def onupdate(self, module, resource, obj):
-        self.onload(module, resource, obj)
+    def on_update(self, module, resource, obj):
+        self.on_load(module, resource, obj)
 
-    def ondelete(self, module, name, value):
+    def on_delete(self, module, name, value):
         by_module_resource[module, name].close()
         del by_module_resource[module, name]
         if lookup:
             lookup.invalidate_cache()
 
-    def oncreaterequest(self, module, name, kwargs):
+    def on_create_request(self, module, name, kwargs):
         resourceobj = {
             "resource_type": self.type,
             "folder": kwargs["folder"],
@@ -70,7 +70,7 @@ class FileServerType(modules_state.ResourceType):
 
         return resourceobj
 
-    def onupdaterequest(self, module, resource, resourceobj, kwargs):
+    def on_update_request(self, module, resource, resourceobj, kwargs):
         resourceobj = {
             "resource_type": self.type,
             "folder": kwargs["folder"],
@@ -86,7 +86,7 @@ class FileServerType(modules_state.ResourceType):
 
         return resourceobj
 
-    def createpage(self, module, path):
+    def create_page(self, module, path):
         d = SimpleDialog(f"New File Server in {module}")
 
         d.text("File servers can be accessed at /pages/modulename/resourcename/file/name.")
@@ -102,7 +102,7 @@ class FileServerType(modules_state.ResourceType):
         d.submit_button("Create")
         return d.render(f"/modules/module/{url(module)}/addresourcetarget/{self.type}", hidden_inputs={"path": path})
 
-    def editpage(self, module, resource, resource_data):
+    def edit_page(self, module, resource, resource_data):
         if "require_permissions" in resource_data:
             requiredpermissions = resource_data["require_permissions"]
         else: